==> R/02_pathogen_data_cleaning.Rmd <==
---
title: "BioMove swallow-parasite manuscript (2)" ## name of your project and analysis step
description: "cleaning parasite data"
author:
    - name: "Marius Grabow"
      email: "grabow@izw-berlin.de" 
      affiliation: Leibniz Institute for Zoo and Wildlife Research
      affiliation_url: https://izw-berlin.de/en/
      orcid_id: 0000-0003-0702-9642
date: "`r Sys.Date()`"
output: 
    distill::distill_article:
        highlight: kate      ## styling of code
        code_folding: true  ## if `true` you can expand and shrink code chunks
        toc: true            ## if `true` adds a table of content
        toc_depth: 2         ## level to be displayed in the table of content
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, 
                      dev = "ragg_png", fig.width = 9, fig.height = 6, dpi = 600, retina = 1)
Sys.setlocale("LC_TIME", "C")
```

# Setup and R-packages needed for analysis

```{r packages}
library("tidyverse")
library("here")
library("lubridate")
```

# Reading in our data set from script 01: Capture-mark-Recapture data was filtered for data on swallows only

```{r data}
cmr <- readRDS(here("output",
                    "data-proc",
                    "cmr_swallows.RDS"))
```
# Clean the infection data
```{r}
infection_19 <- read_csv(here("data-raw",
                              "pathogen_data",
                              "infection_blood_2019.csv")) %>%
  select(date, ring_id, plasmodium, haemoproteus, leucocytozoon) %>%
  filter(!is.na(ring_id)) %>%
  mutate(
    date = dmy(date),
    year = year(date),
    year = ifelse(year != 2019, 2019, year)
  )



infection_20 <- read_csv(here("data-raw",
                              "pathogen_data",
                              "infection_blood_2020.csv")) %>%
  select(date, ring_id, plasmodium, haemoproteus, leucocytozoon) %>%
  filter(!is.na(ring_id)) %>%
  mutate(
    date = dmy(date),
    year = year(date),
    year = ifelse(year != 2020, 2020, year)
  )

infection_21 <- read_csv(here("data-raw",
                              "pathogen_data",
                              "infection_blood_2021.csv")) %>%
  select(date, ring_id, plasmodium, haemoproteus, leucocytozoon) %>%
  filter(!is.na(ring_id)) %>%
  mutate(
    date = dmy(date),
    year = year(date),
    year = ifelse(year != 2021, 2021, year)
  )

infection_22 <- read_csv(here("data-raw",
                              "pathogen_data",
                              "infection_blood_2022.csv")) %>%
  select(date, ring_id, plasmodium, haemoproteus, leucocytozoon) %>%
  filter(!is.na(ring_id)) %>%
  mutate(date = dmy(date),
         year = 2022)

infection_23 <- read_csv(here("data-raw",
                              "pathogen_data",
                              "infection_blood_2023.csv")) %>%
  select(date, ring_id, plasmodium, haemoproteus, leucocytozoon) %>%
  filter(!is.na(ring_id)) %>%
  mutate(date = dmy(date),
         year = 2023)

#manipulate infection information to binary:

infection <- infection %>%
  group_by(ring_id) %>%
  mutate(
    i_2020 = case_when(
      blood_infection == "n" & year == 2020 ~ "n",
      blood_infection == "y" & year == 2020 ~ "y"
    ),
    i_2021 = case_when(
      blood_infection == "n" & year == 2021 ~ "n",
      blood_infection == "y" & year == 2021 ~ "y"
    ),
    i_2022 = case_when(
      blood_infection == "n" & year == 2022 ~ "n",
      blood_infection == "y" & year == 2022 ~ "y"
    ),
    i_2023 = case_when(
      blood_infection == "n" & year == 2023 ~ "n",
      blood_infection == "y" & year == 2023 ~ "y"
    ))%>%
  fill(i_2020:i_2023,.direction = "updown")%>%
  ungroup()

```

# messed up locations from field sheets, correcting here manually to be comparable to Wiebkes study on the swallows

```{r}
cmr_new <- cmr %>%
  mutate(
    location = case_when(
      location %in% c("Lutz N",
                      "Lutz Nix",
                      "Lutz-Nix",
                      "Lutz Dedelow",
                      "Dedelow Mahlzahn",
                      "Dedelow (Wo?)",
                      "Hackbart") ~ "Dedelow",

      location %in% c("Dedelow-RS",
                      "Dedelow-MS",
                      "Wiebke",
                      "Futterhaus",
                      "Mehlis II MVA") ~ "Dedelow-MVA",

      location %in% c("MVA-Kutzerow",
                      "Menke-Kutzerow") ~ "MVA-Kutzerow",

      location %in% c("Holzendorf (schwanger. 2 Hunde)",
                      "Holztendorf M\xfcller",
                      "Holzendorf M\xfcller",
                      "Fom.Martin Holzendorf") ~ "Martin",

      location %in% c("Falkenhagen (wo?)",
                      "Falkenhagen (18)",
                      "Falkenhagen (6)",
                      "Falkenhagen",
                      "Falkenhagen Kin",
                      "PRIVAT, Labes, Falkenhagen") ~ "Labes",

      location %in% c("Rens",
                      "R\xe4nsch Falkenhagen") ~ "Rens")
  )
```


# Join pathogen data with CMR
## Save data for further analysis
## Calculate prevalences
```{r}

Swallow_infection <- left_join(cmr_new,
                               infection %>% select(-date),
                               by = c("ring_id", "year"))

saveRDS(Swallow_infection,
        here("output",
             "data-proc",
             "cmr_pathogens.RDS"))

```

***

<details><summary>Session Info</summary>

```{r sessionInfo}
## DO NOT REMOVE!
## We store the settings of your computer and the current versions of the
## packages used to allow for reproducibility
Sys.time()
#git2r::repository() ## uncomment if you are using GitHub
sessionInfo()
```

</details>
