==> R/03_movment_data_cleaning.Rmd <==
---
title: "BioMove swallow-parasite manuscript (3)" ## name of your project and analysis step
description: "cleaning raw movement data"
author:
    - name: "Marius Grabow"
      email: "grabow@izw-berlin.de" 
      affiliation: Leibniz Institute for Zoo and Wildlife Research
      affiliation_url: https://izw-berlin.de/en/
      orcid_id: 0000-0003-0702-9642
date: "`r Sys.Date()`"
output: 
    distill::distill_article:
        highlight: kate      ## styling of code
        code_folding: true  ## if `true` you can expand and shrink code chunks
        toc: true            ## if `true` adds a table of content
        toc_depth: 2         ## level to be displayed in the table of content
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, 
                      dev = "ragg_png", fig.width = 9, fig.height = 6, dpi = 600, retina = 1)
Sys.setlocale("LC_TIME", "C")
```

# Set-up

```{r message=FALSE, warning=FALSE}
package.list = c(
  "tidyverse",
  "here",
  "tmap",
  "sf",
  "lubridate"
)

for (package in package.list) {
  if (!require(package, character.only = T, quietly = T)) {
    install.packages(package)
    library(package, character.only = T)
  }
}
```

# Read data base (only working with VPN)
```{r}
df<-read_csv(here("data-raw",
                 "swallows_filtered.csv"))
```
# Get CMR data
## This data set contains all information about the single individuals captured (body measurements, sex, etc.), but also information of the infection status for years that have already been analysed
```{r}
###cmr data for filtering
cmr <- readRDS(here("output", "data-proc", "cmr_pathogens.RDS")) %>%
  mutate(tag_id = as.factor(tag_id)) %>%
  dplyr::select(
    date,
    tag_id,
    ring_id,
    location,
    species,
    sex,
    blood_infection,
    sum_infection,
    plasmodium,
    haemoproteus,
    leucocytozoon,
    infection_type,
    capture_type
  ) %>%
  mutate(date = as.Date(date))
```


# Spatial analyses & data manipulation
```{r}
swallows_df <- df %>%
  #shorten tag-id to 4 digits
  mutate(TAG = as.factor(substr(TAG, 8, 11))) %>%
  # human-readable time
  mutate(TIME = as.POSIXct(((TIME + 0.1) / 1000), origin = "1970-01-01", tz =
                             "UTC")) %>%
  mutate(TIME = ymd_hms(TIME)) %>%
  mutate(hms = hms(substr(TIME, 11, 19))) %>%
  group_by(TAG) %>%
  # time differences
  mutate(difftime = abs(difftime(TIME, dplyr::lag(TIME), units = "secs"))) %>%
  ungroup() %>%
  # date
  mutate(date = as.Date(substr(TIME, 1, 10))) %>%
  group_by(TAG) %>%
  # tagtime
  mutate(tagtime_sum = max(TIME) - min(TIME),
         tagtime = TIME - min(TIME)) %>%
  ungroup() %>%
  arrange(TAG, TIME) %>%
  #spatial
  st_as_sf(coords = c("X", "Y"),
           crs = 32633,
           remove = FALSE) %>%
  rename(tag_id = TAG) %>%
  st_transform(crs = 3035) %>%
  # lat long coords
  mutate(long = st_coordinates(st_transform(., 4326))[, 1],
         lat = st_coordinates(st_transform(., 4326))[, 2])

swallows <-
  left_join(swallows_df, cmr %>% dplyr::select(-date), by = c("tag_id"))
```

# Biological filtering (exclude all observations during night)
## suncalc package gives multiple options, how the night ends, I choose civil twilight instead of sunrise, because bird activity before the tip of the sun is actually visible (i.e. sunrise).


```{r}
### speed and covariance filter----
swallows_filter <- swallows %>%
  #filter out all observations with high covariance
  filter(abs(COVXY) < 50 * 50) %>%
  group_by(tag_id) %>%
  #recalculate distances,difftime & speed after some observations have been removed
  mutate(
    distance_from_last = st_distance(lag(geometry),
                                     geometry, by_element = T),
    difftime = abs(difftime(TIME, lag(TIME), units = "secs")),
    speed_estimate_ms = as.numeric(distance_from_last) / as.numeric(difftime)
  ) %>%
  ungroup() %>%
  #filter out high speed estimates
  filter(speed_estimate_ms <= 30)

### biological filtering (exclude all observations at night, here defined as dawn / nautical dusk)

sunlight_time <- suncalc::getSunlightTimes(
  date = as.Date(swallows_filter$date),
  lat = mean(swallows_filter$lat),
  lon = mean(swallows_filter$long),
  keep = c("dawn", "nauticalDusk"),
  tz = "GMT"
) %>%
  mutate(dawn = ymd_hms(dawn)) %>%
  mutate(nauticalDusk = ymd_hms(nauticalDusk)) %>%
  mutate(interval = interval(dawn, nauticalDusk))

swallows_filter_day <-
  swallows_filter %>%
  tibble::add_column(sunlight_time$interval) %>%
  #keep just daylight observations
  filter(TIME %within% sunlight_time$interval) %>%
  # drop geometry columns
  st_drop_geometry()


saveRDS(swallows_filter_day,
        here("output",
             "data-proc",
             "swallows_day_all.RDS"))

```

***

<details><summary>Session Info</summary>

```{r sessionInfo}
## DO NOT REMOVE!
## We store the settings of your computer and the current versions of the
## packages used to allow for reproducibility
Sys.time()
#git2r::repository() ## uncomment if you are using GitHub
sessionInfo()
```

</details>
