==> R/06_ISSF_habitat_selection_models.Rmd <==
---
title: "BioMove swallow-parasite manuscript (6)" ## name of your project and analysis step
description: "Integrated Step Selection function - microhabitat selection"
author:
    - name: "Marius Grabow"
      email: "grabow@izw-berlin.de" 
      affiliation: Leibniz Institute for Zoo and Wildlife Research
      affiliation_url: https://izw-berlin.de/en/
      orcid_id: 0000-0003-0702-9642
date: "`r Sys.Date()`"
output: 
    distill::distill_article:
        highlight: kate      ## styling of code
        code_folding: true  ## if `true` you can expand and shrink code chunks
        toc: true            ## if `true` adds a table of content
        toc_depth: 2         ## level to be displayed in the table of content
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, 
                      dev = "ragg_png", fig.width = 9, fig.height = 6, dpi = 600, retina = 1)
Sys.setlocale("LC_TIME", "C")
```

# Setup

```{r packages}
library("tidyverse")
library("terra")
library("amt")
library("sf")
library("here")
library("broom")
```
# Load data
```{r}
MS<-readRDS(here("output",
             "data-proc",
             "MS_states_X.RDS"))

RS<-readRDS(here("output",
             "data-proc",
             "RS_states_X.RDS"))

swallows<-bind_rows(MS,RS)%>%
  group_by(tag_id,species,blood_infection)%>%
  summarise()%>%
  group_by(species,blood_infection)%>%
  summarise(n=n())

# environmental raster
ras<-terra::rast(here("output",
                  "geo-proc",
                  "stacked.tif"))


# create binary raster for water bodies, adding a 25m buffer around all waterbodies to include the habitat around it
binary_raster_w <- ifel(ras$dist_water >= 0 & ras$dist_water <= 25, 4, 0)

# reclassify
ras$landuse_reclass<-ifel(binary_raster_w == 4, 4, ras$landuse_reclass)

# plot reclassified raster
plot(ras$landuse_reclass)

cmr_pathogens<-readRDS(here("output",
                            "data-proc",
                            "cmr_pathogens.RDS"))%>%
  filter(!is.na(blood_infection))%>%
  select(tag_id,blood_infection)%>%
  mutate(blood_infection=as.factor(blood_infection))%>%
  dplyr::mutate(tag_id=as.factor(tag_id))

```
# Make track xyt   
## MS
```{r, MS track}

MS_track <- MS %>%
  as.data.frame() %>%
  #filter(state!=1)%>%
  #make amt track
  make_track(x,
             y,
             TIME,
             crs = 3035,
             all_cols = TRUE,
             id = tag_id) %>%
  nest(data = -c(tag_id, blood_infection)) %>%
  mutate(data = map(data, ~ track_resample(
    .x,
    rate = minutes(1),
    tolerance = seconds(10)
  ))) %>%
  mutate(
    data = map(data, ~ steps_by_burst(.x, keep_cols = "end")),
    data = map(data, ~ random_steps(.x, ncontrol = 5)),
    data = map(data, ~ extract_covariates(.x, ras)),
    data = map(
      data,
      ~ mutate(
        .x,
        log_sl_ = log(sl_),
        sin_ta_ = sin(ta_),
        step_id1_ = paste(tag_id, step_id_, sep =
                            "."),
        landuse_reclass = as.factor(landuse_reclass)
      )
    )
  )
```

## Model 3: Random Intercept + random slope

```{r}
dat.issf.m3 <- MS_track %>%
  unnest(cols = data) %>%
  mutate(
    human1 = ifelse(landuse_reclass == 1, 1, 0),
    agriculture1 = ifelse(landuse_reclass == 2, 1, 0),
    natural1 = ifelse(landuse_reclass == 3, 1, 0),
    water1 = ifelse(landuse_reclass == 4, 1, 0)
  )


MS_m3 <-
  glmmTMB::glmmTMB(
    case_ ~ -1 + agriculture1 * blood_infection + natural1 * blood_infection + water1 * blood_infection + dist_human * blood_infection +
      sl_ +  (1 |
                tag_id) + (0 + agriculture1 |
                             tag_id) + (0 + natural1 |
                                          tag_id) + (0 + water1 | tag_id),
    family = poisson,
    data = dat.issf.m3,
    doFit = FALSE
  )

MS_m3$theta[1] <- log(1e3)
MS_m3$mapArg <- list(theta = factor(c(NA, 1:3)))

MS_m3 <- glmmTMB::fitTMB(MS_m3)

summary(MS_m3)

tidy_MS_m3 <- broom.mixed::tidy(MS_m3) %>%
  filter(effect == "fixed") %>%
  mutate(
    habitat = case_when(
      str_detect(term, "agriculture") ~ "agriculture",
      str_detect(term, "natural") ~ "natural",
      str_detect(term, "water") ~ "water",
      TRUE ~ "other"  # default case
    )
  ) %>%
  filter(habitat != "other") %>%
  mutate(blood_infection = case_when(str_detect(term, "infectiony") ~ "yes",
                                     TRUE ~ "no")) # default case
                                     
                                     
# plot                                  
tidy_MS_m3 %>%
  ggplot(aes(
    x = habitat,
    y = estimate,
    colour = blood_infection,
    group = blood_infection
  )) +
  geom_pointrange(
    aes(
      ymin = estimate - 1.96 * std.error,
      ymax = estimate + 1.96 * std.error
    ),
    position = position_jitterdodge(jitter.width = 0.05),
    size = 2,
    linewidth = 3
  ) +
  d6::theme_d6(base_size = 30) +
  ylim(-0.6, 0.6) +
  labs(y = "Relative selection strength") +
  scale_colour_manual(
    values = c("#5BBCD6", "#F98400"),
    labels = c("non-infected", "infected"),
    name = "Infection status"
  )
                                     

```

## Model 4: Random Intercept + random slope (just foraging)

```{r}
dat.issf.m4<-MS%>%
  as.data.frame()%>%
  filter(state==2)%>%
  #make amt track
  make_track(x,y,TIME,
             crs = 3035,
             all_cols=TRUE,
             id=tag_id)%>%
  nest(data= -c(tag_id,blood_infection))%>%
  mutate(data = map(
    data, ~ track_resample(.x,
                           rate=minutes(1),
                           tolerance=seconds(10))))%>%
  mutate(data = map(data, ~steps_by_burst(.x,keep_cols="end")),
         data = map(data, ~random_steps(.x,ncontrol=5)),
         data = map(data, ~extract_covariates(.x,ras)),
         data = map(data, ~mutate(.x,
                                  log_sl_ = log(sl_),
                                  sin_ta_ = sin(ta_),
                                  step_id1_=paste(tag_id,step_id_,sep="."),
                                  landuse_reclass=as.factor(landuse_reclass))))%>%
  unnest(cols = data)%>%
  mutate(human1=ifelse(landuse_reclass==1,1,0),
         agriculture1=ifelse(landuse_reclass==2,1,0),
         natural1=ifelse(landuse_reclass==3,1,0),
         water1=ifelse(landuse_reclass==4,1,0))


MS_m4 <-
  glmmTMB::glmmTMB(
    case_ ~ -1 + agriculture1 * blood_infection + natural1 * blood_infection + water1 * blood_infection + dist_human * blood_infection +
      sl_ +  (1 |
                tag_id) + (0 + agriculture1 |
                             tag_id) + (0 + natural1 | tag_id) + (0 + water1 | tag_id),
    family = poisson,
    data = dat.issf.m4,
    doFit = FALSE
  )

MS_m4$theta[1] <- log(1e3)
MS_m4$mapArg <-list(theta=factor(c(NA,1:3)))

MS_m4<-glmmTMB::fitTMB(MS_m4)

summary(MS_m4)

tidy_MS_m4<-broom.mixed::tidy(MS_m4)%>%
  filter(effect=="fixed")%>%
  mutate(habitat = case_when(
    str_detect(term, "agriculture") ~ "agricultural",
    str_detect(term, "natural") ~ "semi-natural",
    str_detect(term, "water") ~ "water-related",
    TRUE ~ "other"  # default case 
  ))%>%
  filter(habitat!="other")%>%
  mutate(blood_infection = case_when(
    str_detect(term, "infectiony") ~ "yes",
    TRUE ~ "no"  # default case 
  ))


tidy_MS_m4 %>%
  ggplot(aes(
    x = habitat,
    y = estimate,
    colour = blood_infection,
    group = blood_infection
  )) +
  geom_pointrange(
    aes(
      ymin = estimate - 1.96 * std.error,
      ymax = estimate + 1.96 * std.error
    ),
    position = position_jitterdodge(jitter.width = 0.05),
    size = 2,
    linewidth = 3
  ) +
  d6::theme_d6(base_size = 35) +
  ylim(-1, 1) +
  labs(y = "Relative selection strength",
       x = "Habitat-type") +
  geom_hline(
    yintercept = 0,
    lty = "dashed",
    colour = "grey20",
    linewidth = 1.5
  ) +
  scale_colour_manual(
    values = c("#5BBCD6", "#F98400"),
    labels = c("non-infected", "infected"),
    name = "Infection status"
  )


```

# Make track xyt
## RS
```{r, RS}

RS_track<-RS%>%
  as.data.frame()%>%
  #filter(state!=1)%>%
  #make amt track
  make_track(x,y,TIME,
             crs = 3035,
             all_cols=TRUE,
             id=tag_id)%>%
  nest(data= -c(tag_id,blood_infection))%>%
  mutate(data = map(
    data, ~ track_resample(.x,
                           rate=minutes(1),
                           tolerance=seconds(10))))%>%
  mutate(data = map(data, ~steps_by_burst(.x,keep_cols="end")),
         data = map(data, ~random_steps(.x,ncontrol=5)),
         data = map(data, ~extract_covariates(.x,ras)),
         data = map(data, ~mutate(.x,
                                  log_sl_ = log(sl_),
                                  sin_ta_ = sin(ta_),
                                  step_id1_=paste(tag_id,step_id_,sep="."),
                                  landuse_reclass=as.factor(landuse_reclass))))
```

## Model 3: Random Intercept + random slope

```{r}
RS_dat.issf.m3<-RS_track%>%
  unnest(cols = data)%>%
  mutate(human1=ifelse(landuse_reclass==1,1,0),
         agriculture1=ifelse(landuse_reclass==2,1,0),
         natural1=ifelse(landuse_reclass==3,1,0),
         water1=ifelse(landuse_reclass==4,1,0))


RS_m3 <-
  glmmTMB::glmmTMB(
    case_ ~ -1 + agriculture1 * blood_infection + natural1 * blood_infection + water1 * blood_infection + dist_human * blood_infection +
      sl_ +  (1 |
                tag_id) + (0 + agriculture1 |
                             tag_id) + (0 + natural1 | tag_id) + (0 + water1 | tag_id),
    family = poisson,
    data = RS_dat.issf.m3,
    doFit = FALSE
  )

RS_m3$theta[1] <- log(1e3)
RS_m3$mapArg <-list(theta=factor(c(NA,1:3)))

RS_m3<-glmmTMB::fitTMB(RS_m3)

summary(RS_m3)

tidy_RS_m3<-broom.mixed::tidy(RS_m3)%>%
  filter(effect=="fixed")%>%
  mutate(habitat = case_when(
    str_detect(term, "agriculture") ~ "agriculture",
    str_detect(term, "natural") ~ "semi-natural",
    str_detect(term, "water") ~ "water",
    TRUE ~ "other"  # default case 
  ))%>%
  filter(habitat!="other")%>%
  mutate(blood_infection = case_when(
    str_detect(term, "infectiony") ~ "yes",
    TRUE ~ "no"  # default case 
  ))

tidy_RS_m3 %>%
  ggplot(aes(
    x = habitat,
    y = estimate,
    colour = blood_infection,
    group = blood_infection
  )) +
  geom_pointrange(
    aes(
      ymin = estimate - 1.96 * std.error,
      ymax = estimate + 1.96 * std.error
    ),
    position = position_jitterdodge(jitter.width = 0.05),
    size = 2,
    linewidth = 3
  ) +
  d6::theme_d6(base_size = 30) +
  geom_hline(yintercept = 0,
             lty = "dashed",
             colour = "grey20") +
  ylim(-0.6, 0.6) +
  labs(y = "Relative selection strength") +
  scale_colour_manual(
    values = c("#5BBCD6", "#F98400"),
    labels = c("non-infected", "infected"),
    name = "Infection status"
  )


glmmTMB::ranef(RS_m3)[[1]]

```

## Model 4: Random Intercept + random slope (just foraging)

```{r}
RS.dat.issf.m4<-RS%>%
  as.data.frame()%>%
  filter(state==2)%>%
  #make amt track
  make_track(x,y,TIME,
             crs = 3035,
             all_cols=TRUE,
             id=tag_id)%>%
  nest(data= -c(tag_id,blood_infection))%>%
  mutate(data = map(
    data, ~ track_resample(.x,
                           rate=minutes(1),
                           tolerance=seconds(10))))%>%
  mutate(data = map(data, ~steps_by_burst(.x,keep_cols="end")),
         data = map(data, ~random_steps(.x,ncontrol=5)),
         data = map(data, ~extract_covariates(.x,ras)),
         data = map(data, ~mutate(.x,
                                  log_sl_ = log(sl_),
                                  sin_ta_ = sin(ta_),
                                  step_id1_=paste(tag_id,step_id_,sep="."),
                                  landuse_reclass=as.factor(landuse_reclass))))%>%
  unnest(cols = data)%>%
  mutate(human1=ifelse(landuse_reclass==1,1,0),
         agriculture1=ifelse(landuse_reclass==2,1,0),
         natural1=ifelse(landuse_reclass==3,1,0),
         water1=ifelse(landuse_reclass==4,1,0))


RS_m4 <-
  glmmTMB::glmmTMB(
    case_ ~ -1 + agriculture1 * blood_infection + natural1 * blood_infection + water1 * blood_infection + dist_human * blood_infection +
      sl_ +  (1 |
                tag_id) + (0 + agriculture1 |
                             tag_id) + (0 + natural1 | tag_id) + (0 + water1 | tag_id),
    family = poisson,
    data = RS.dat.issf.m4,
    doFit = FALSE
  )

RS_m4$theta[1] <- log(1e3)
RS_m4$mapArg <-list(theta=factor(c(NA,1:3)))

RS_m4<-glmmTMB::fitTMB(RS_m4)
summary(RS_m4)

# tidy model results
tidy_RS_m4 <- broom.mixed::tidy(RS_m4) %>%
  filter(effect == "fixed") %>%
  mutate(
    habitat = case_when(
      str_detect(term, "agriculture") ~ "agricultural",
      str_detect(term, "natural") ~ "semi-natural",
      str_detect(term, "water") ~ "water-related",
      TRUE ~ "other"  # default case
    )
  ) %>%
  filter(habitat != "other") %>%
  mutate(blood_infection = case_when(str_detect(term, "infectiony") ~ "yes",
                                     TRUE ~ "no"))

# plot
tidy_RS_m4 %>%
  ggplot(aes(
    x = habitat,
    y = estimate,
    colour = blood_infection,
    group = blood_infection
  )) +
  geom_pointrange(
    aes(
      ymin = estimate - 1.96 * std.error,
      ymax = estimate + 1.96 * std.error
    ),
    position = position_jitterdodge(jitter.width = 0.05),
    size = 2,
    linewidth = 3
  ) +
  d6::theme_d6(base_size = 35) +
  geom_hline(
    yintercept = 0,
    lty = "dashed",
    colour = "grey20",
    linewidth = 1.5
  ) +
  ylim(-1, 1) +
  labs(y = "Relative selection strength",
       x = "Habitat-type") +
  scale_colour_manual(
    values = c("#5BBCD6", "#F98400"),
    labels = c("non-infected", "infected"),
    name = "Infection status"
  )

```
