==> R/07_multivaraite_analyses.Rmd <==
---
title: "BioMove swallow-parasite manuscript (8)" ## name of your project and analysis step
description: "Multivariate analyses"
author:
    - name: "Marius Grabow"
      email: "grabow@izw-berlin.de" 
      affiliation: Leibniz Institute for Zoo and Wildlife Research
      affiliation_url: https://izw-berlin.de/en/
      orcid_id: 0000-0003-0702-9642
date: "`r Sys.Date()`"
output: 
    distill::distill_article:
        highlight: kate      ## styling of code
        code_folding: true  ## if `true` you can expand and shrink code chunks
        toc: true            ## if `true` adds a table of content
        toc_depth: 2         ## level to be displayed in the table of content
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library("tidyverse")
library("here")
library("vegan")
library("factoextra")
library("patchwork")

```
# Data
```{r}
swallows <- readRDS(here("output",
                         "data-proc",
                         "cmr_pathogens.RDS"))
# fill traits of recaptures
traits <- swallows %>%
  mutate(year = lubridate::year(date)) %>%
  select(
    c(
      "ring_id",
      "year",
      "species",
      "age",
      "sex",
      "tars_mm",
      "wing_mm",
      "X8prim_mm",
      "fat",
      "muscle",
      "tail_mm",
      "tail_fork_mm",
      "weight_g",
      "min_age",
      "catch_nr_all_time",
      "blood_infection",
      "plasmodium",
      "leucocytozoon",
      "haemoproteus"
    )
  ) %>%
  group_by(ring_id, year) %>%
  tidyr::fill(.direction = "downup") %>%
  group_by(ring_id) %>%
  tidyr::fill(.direction = "downup") %>%
  mutate(body_condition = weight_g / tars_mm) %>%
  mutate(
    blood_infection = case_when(
      blood_infection == "y" ~ "infected",
      blood_infection == "n" ~ "non-infected"
    )
  )
# split per species
## barn swallow
RS_traits <- traits %>%
  filter(species == "RS") %>%
  ungroup() %>%
  dplyr::select(
    year,
    sex,
    tars_mm,
    wing_mm,
    fat,
    muscle,
    tail_fork_mm,
    weight_g,
    min_age,
    catch_nr_all_time,
    body_condition,
    blood_infection
  ) %>%
  filter(!is.na(blood_infection)) %>%
  na.omit()
## hosue martin
MS_traits <- traits %>%
  filter(species == "MS") %>%
  ungroup() %>%
  dplyr::select(
    year,
    sex,
    tars_mm,
    wing_mm,
    fat,
    muscle,
    weight_g,
    min_age,
    catch_nr_all_time,
    body_condition,
    blood_infection
  ) %>%
  filter(!is.na(blood_infection)) %>%
  na.omit()

```
# PCA: Barn swallow
```{r}
res.pca <-
  prcomp(RS_traits %>% select(tars_mm:body_condition), scale = TRUE)
#Eigenvector
fviz_eig(res.pca)
# PCA graph of individuals
fviz_pca_ind(
  res.pca,
  col.ind = "cos2",
  # Color by the quality of representation
  gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"),
  repel = TRUE     # Avoid text overlapping
)

# PCA graph of variables
fviz_pca_var(
  res.pca,
  col.var = "contrib",
  # Color by contributions to the PC
  gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"),
  repel = TRUE     # Avoid text overlapping
)

#Biplot of individuals and variables
fviz_pca_biplot(
  res.pca,
  repel = TRUE,
  col.var = "#2E9FDF",
  # Variables color
  col.ind = "#696969"  # Individuals color
)

# group by blood infection
groups <- as.factor(RS_traits$blood_infection)
# biplot with groups
(
  p1 <- fviz_pca_biplot(
    res.pca,
    col.ind = groups,
    # color by groups
    palette = c("#F98400", "#5BBCD6"),
    alpha.ind = 1,
    addEllipses = TRUE,
    # Concentration ellipses
    ellipse.type = "confidence",
    legend.title = "Blood infection",
    repel = TRUE,
    pointsize = 4,
    col.var = "red",
    labelsize = 10,
    textsize = 10,
    label = "var"
  ) +
    labs(title = element_blank()) +
    d6::theme_d6(base_size = 30,
                 legend = "top") +
    theme(plot.background = element_rect(fill = "white"))
)

```
# House martin
```{r}


MS.pca <-
  prcomp(MS_traits %>% select(tars_mm:body_condition), scale = TRUE)

# PCA graph of variables
fviz_pca_var(
  MS.pca,
  col.var = "contrib",
  # Color by contributions to the PC
  gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"),
  repel = TRUE     # Avoid text overlapping
)
# group by infection status
MS_groups <- as.factor(MS_traits$blood_infection)

#biplot pca
(
  p2 <- fviz_pca_biplot(
    MS.pca,
    col.ind = MS_groups,
    # color by groups
    palette = c("#F98400", "#5BBCD6"),
    alpha.ind = 0.5,
    addEllipses = TRUE,
    # Concentration ellipses
    ellipse.type = "confidence",
    legend.title = "Blood infection",
    repel = TRUE,
    pointsize = 4,
    col.var = "red",
    labelsize = 10,
    textsize = 10,
    label = "var"
  ) +
    labs(title = element_blank()) +
    d6::theme_d6(base_size = 30,
                 legend = "top") +
    theme(plot.background = element_rect(fill = "white"))
)

```
# combined plot
```{r}
p1 + p2 + plot_layout(guides = "collect") &
  theme(legend.position = 'bottom')

```
