==> R/10_resampling_30min.Rmd <==
---
title: "XXX_test_sampling"
output: html_document
date: "2024-02-14"
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library("here")
library("tidyverse")
library("momentuHMM")
library("sf")
library("ctmm")

prep<-function(df,TIME){
  
  
  df%>%
  sf::st_drop_geometry() %>%
  # filter out all animals that have not been blood-sampled
  filter(!is.na(blood_infection)) %>%
  #rename
  mutate(
    sensor = "atlas",
    tag_id = as.character(tag_id),
    TAG = tag_id) %>%
  #doesn't like tibbles
  as.data.frame() %>%
  #rename
  rename(
    individual.local.identifier = tag_id,
    timestamp = TIME,
    location.long = long,
    location.lat = lat) %>%
  #telemetry object
  as.telemetry(
    projection = "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs +type=crs",
    keep = c(
      "TAG",
      "species",
      "sex",
      "blood_infection")
  )
}
```
# 30 min
```{r}
swallows8<-readRDS(here("output",
             "data-proc",
             "swallows_final_2024.RDS"))%>%
  ungroup()%>%
  st_as_sf(coords=c("x","y"),crs=3035,remove=F)


cmr<-readRDS(here("output",
                  "data-proc",
                  "cmr_pathogens.RDS"))%>%
  dplyr::select(tag_id,blood_infection,species,location
         )%>%
  mutate(tag_id=as.factor(tag_id))%>%
  rename(capture_location=location)%>%
  distinct(tag_id,.keep_all = T)
  


swallows_30min<-swallows8%>%
  group_by(tag_id)%>% 
  mutate(row_within_group = row_number()) %>% 
  # test 30 minutes * 60 seconds / 8 seconds = 225
  filter(row_within_group %% 225 == 0) %>%
  select(-row_within_group)%>%
  ungroup()%>%
  st_as_sf(coords=c("x","y"),crs=3035,remove=F)

d6atlas::atlasplot()+
  geom_sf(data=swallows8,col="red")+
  geom_sf(data=swallows_30min,col="blue")


s_combined<-bind_rows(swallows8%>%mutate(resolution="8 sec"),
                      swallows_30min%>%mutate(resolution="30 min"))


d6atlas::atlasplot()+
  geom_sf(data=s_combined,aes(col=resolution))+
  d6::theme_d6()



ggsave(here("plots",
            "compare_8sec_30min.tiff"))



swallows_10min<-swallows8%>%
  group_by(tag_id)%>% 
  mutate(row_within_group = row_number()) %>% 
  # test 30 minutes * 60 seconds / 8 seconds = 225
  filter(row_within_group %% 75 == 0) %>%
  select(-row_within_group)%>%
  ungroup()%>%
  st_as_sf(coords=c("x","y"),crs=3035,remove=F)
# 
# d6atlas::atlasplot()+
#   geom_sf(data=swallows8,col="red")+
#   geom_sf(data=swallows_30min,col="blue")
# 
# ggsave(here("plots",
#             "compare_8sec_30min.tiff"))


```
#ctmm 30 min
```{r}

swallows<-left_join(swallows_30min,cmr,by="tag_id")%>%
  mutate(blood_infection=as.factor(blood_infection))%>%
  mutate(date=as_date(TIME))%>%
  unite("ID",tag_id,date,remove=F)


ctmm_prep<-swallows%>%
  sf::st_drop_geometry() %>%
  # filter out all animals that have not been blood-sampled
  filter(!is.na(blood_infection)) %>%
  #rename
  mutate(
    sensor = "atlas",
    tag_id = as.character(tag_id),
    TAG = tag_id) %>%
  #doesn't like tibbles
  as.data.frame() %>%
  #rename
  rename(
    individual.local.identifier = tag_id,
    timestamp = TIME,
    location.long = long,
    location.lat = lat)%>%
  #telemetry object
  as.telemetry(
    projection = "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs +type=crs",keep = TRUE)


uere(ctmm_prep)<-1


# set-up parallel sessions
cl <- parallel::makeCluster(4)
doParallel::registerDoParallel(cl)

# create function to fit data & select best movement model
model_fit <- function(i)
{
  GUESS <-
    ctmm.guess(ctmm_prep[[i]],
               CTMM = ctmm(error = TRUE),
               interactive = FALSE)
  ctmm.select(ctmm_prep[[i]], GUESS)
}

#iterate over list
ctmm_fit_30min <-
  foreach(
    i = 1:length(ctmm_prep),
    .packages = 'ctmm',
    .errorhandling = 'remove'
  ) %dopar% {
    model_fit(i)
  }

#same names
names(ctmm_fit_30min) <- names(ctmm_prep)

# save
saveRDS(ctmm_fit_30min,
        here("output",
             "data-proc",
             "ctmm_fit_30min.RDS"))


# function for calculating akde for each individual (using the ctmm model)
akde_fit_weights <- function(i)
{
  AKDE <- akde(ctmm_prep[[i]],
               ctmm_fit_30min[[i]],
               weights = F)
}

#iterate over list
AKDES_tag_weights <- foreach(i=1: length(ctmm_prep),.packages='ctmm',.errorhandling = 'remove') %dopar% { akde_fit_weights(i) }

names(AKDES_tag_weights)<-names(ctmm_prep)

saveRDS(AKDES_tag_weights,here("output",
                          "data-proc",
                          "akdes_tag_30min.RDS"))

AKDE_size<-NULL

for(i in 1:length(AKDES_tag_weights)){
  TAG <- names(AKDES_tag_weights[i])
  min_AREA <- summary(AKDES_tag_weights[[i]],units=F)$CI[1]
  mean_AREA <- summary(AKDES_tag_weights[[i]],units=F)$CI[2]
  max_AREA <- summary(AKDES_tag_weights[[i]],units=F)$CI[3]
  blood_infection <- ctmm_prep[[i]]$blood_infection[1]
  species <- ctmm_prep[[i]]$species[1]
  
  AKDE_size = rbind(AKDE_size, data.frame(TAG,min_AREA,mean_AREA,max_AREA,blood_infection,species))
  
}


healthy_barn <-  AKDE_size %>%
    mutate(index = row_number())%>%
  filter(species == "RS" & blood_infection == "n")%>%
  pull(index)

infected_barn <-  AKDE_size %>%
    mutate(index = row_number())%>%
  filter(species == "RS" & blood_infection == "y")%>%
  pull(index)

# meta analysis
# akde 95%
meta(list(BS_NEG=
            AKDES_tag_weights[healthy_barn],
          BS_POS=
            AKDES_tag_weights[infected_barn]),
      units=TRUE,
     plot = TRUE,
     verbose = TRUE,
     level.UD = 0.95,
     level = 0.95,debias = F,
     main="homerange 95% AKDE isopleth")


healthy_house <- AKDE_size %>%
    mutate(index = row_number())%>%
  filter(species == "MS" & blood_infection == "n")%>%
  pull(index)

infected_house <- AKDE_size %>%
  mutate(index = row_number()) %>%
  filter(species == "MS" & blood_infection == "y") %>%
  pull(index)

# meta analysis
meta(list(HM_NEG=
            AKDES_tag_weights[healthy_house],
          HM_POS=
            AKDES_tag_weights[infected_house]),
     plot = TRUE,
     verbose = TRUE,
     level.UD = 0.95,
     level = 0.95,debias = F,
     main="homerange 95% AKDE isopleth")

```

# HMM
```{r}

swallows_30min<-left_join(swallows_30min,cmr,by="tag_id")%>%
  mutate(blood_infection=as.factor(blood_infection))%>%
  mutate(date=as_date(TIME))

ras <- terra::rast(here("output",
                        "geo-proc",
                        "stacked_2.tif"))


covariates<-terra::extract(ras,swallows_30min)%>%
  st_drop_geometry()%>%
  select(-ID)%>%
  as.data.frame()


swallows_cov<-bind_cols(swallows_30min,covariates)

RS <- swallows_cov %>% 
  dplyr::filter(species == "RS") %>%
  st_drop_geometry()%>%
  arrange(tag_id,t)%>%
  group_by(tag_id)%>%
  ungroup()%>%
  as.data.frame()%>%
  rename(ID=tag_id)

#prepare data and calculate step lengths and turning angles 
RS_filter<- RS%>%
  group_by(ID)%>%
  mutate(n=n())%>%
  filter(n>50)%>%
  ungroup()%>%as.data.frame()

#prepare data and calculate step lengths and turning angles 

data_hmm_RS <- prepData(data = RS_filter)%>%
  dplyr::mutate(across(.cols = c(dist_kettlehole:dist_river), .fns = ~ as.numeric(scale(.x))))


hist(data_hmm_RS$step,breaks = 20)

hist(data_hmm_RS$angle,breaks = 20)

#Nr states
nbStates<-3
# names
stateNames = c("resting","foraging","commuting")
# distributions
stepDist<-"gamma"
angleDist<-"vm"


# working bounds for step length

stepworkBounds <- matrix(c(rep(-Inf,6),
                           rep(Inf,6)),
                         6,#rows
                         2) #cols

#working bounds for turning angles

angleworkBounds <-  matrix(c(rep(-Inf,6),rep(Inf,6)),6,2)

stepBounds<-matrix(c(0,20,
                     10000,Inf,
                     20000,Inf,#40kmh
                     0,20,
                     0,Inf,
                     0,Inf),nrow=2*nbStates,byrow=TRUE)



angleBounds<-matrix(c(0,pi,
                      0,pi,
                      0,1,
                      0.01,5,
                      0.01,5,
                      2,10),nrow=2*nbStates,byrow=TRUE)


40 * 1000 / 3600 * 60 * 30



stepPar0 <- c(10,500,2000,
              5,200,1000
              #0.95,0.025,0.025
              )

anglePar0 <-c(3.1,3.1,0.01,1,1,1)

Par0_M1<-list(step=stepPar0,
         angle=anglePar0)




RS_2024_X<-fitHMM(data_hmm_RS,
                nbStates=nbStates,
                dist=list(step=stepDist,angle=angleDist),
                Par0=Par0_M1,
                #workBounds=list(step=stepworkBounds,angle=angleworkBounds),
                #userBounds=list(step=stepBounds,angle=angleBounds),
                stateNames=stateNames,
                estAngleMean = list(angle=TRUE),
              formula = ~ dist_kettlehole*blood_infection,
              formulaDelta = ~  blood_infection)



saveRDS(RS_2024_X,here("output",
                    "data-proc",
                    "RS_2024_30min_X.RDS"))



plotStationary(RS_2024_X,plotCI = TRUE)

viterbi_RS_1<-momentuHMM::viterbi(RS_2024_X)

plot(RS_2024_X,plotTracks = T)



Rs_states<-RS_filter%>%
  mutate(state=viterbi_RS_1)%>%
  # not resting
  filter(state!=1)%>%
  #make amt track
  make_track(x,y,TIME,
             crs = 3035,
             all_cols=TRUE,
             id=ID)%>%
  nest(data= -c(ID,blood_infection))%>%
  mutate(data = map(data, ~steps(.x,keep_cols="end")),
         data = map(data, ~random_steps(.x,ncontrol=5)),
         data = map(data, ~extract_covariates(.x,ras)),
         data = map(data, ~mutate(.x,
                                  log_sl_ = log(sl_),
                                  sin_ta_ = sin(ta_),
                                  step_id1_=paste(ID,step_id_,sep="."),
                                  landuse_reclass=as.factor(landuse_reclass))))%>%
  unnest(cols = data)%>%
  mutate(human1=ifelse(landuse_reclass==1,1,0),
         agriculture1=ifelse(landuse_reclass==2,1,0),
         natural1=ifelse(landuse_reclass==3,1,0),
         water1=ifelse(landuse_reclass==4,1,0))


RS_m_30 <-
  glmmTMB::glmmTMB(
    case_ ~ -1 + agriculture1 * blood_infection + natural1 * blood_infection + water1 * blood_infection + dist_human * blood_infection +
      sl_ +  (1 |
                ID) + (0 + agriculture1 |
                             ID) + (0 + natural1 | ID) + (0 + water1 | ID),
    family = poisson,
    data = Ms_states,
    doFit = FALSE
  )

RS_m_30$theta[1] <- log(1e3)
RS_m_30$mapArg <-list(theta=factor(c(NA,1:3)))

RS_m4<-glmmTMB::fitTMB(RS_m_30)

summary(RS_m4)
####


tidy_RS_m4<-broom.mixed::tidy(RS_m4)%>%
  filter(effect=="fixed")%>%
  mutate(habitat = case_when(
    str_detect(term, "agriculture") ~ "agriculture",
    str_detect(term, "natural") ~ "semi-natural",
    str_detect(term, "water") ~ "water",
    TRUE ~ "other"  # default case 
  ))%>%
  filter(habitat!="other")%>%
  mutate(blood_infection = case_when(
    str_detect(term, "infectiony") ~ "yes",
    TRUE ~ "no"  # default case 
  ))



tidy_RS_m4%>%
  ggplot(aes(x=habitat,y=estimate,colour=blood_infection,group=blood_infection))+
  geom_pointrange(aes(ymin = estimate-1.96*std.error, ymax = estimate+1.96*std.error),position = position_jitterdodge(jitter.width = 0.05),size=2,linewidth=3)+
  d6::theme_d6(base_size = 30)+
  ylim(-1,1)+
  labs(y="Relative selection strength")+
  geom_hline(yintercept = 0,lty="dashed",colour="grey20")+
      scale_colour_manual(values = c("#5BBCD6","#F98400"),
                    labels=c("non-infected","infected"),
                    name="Infection status")

#############################

MS <- swallows_cov %>% 
  dplyr::filter(species == "MS") %>%
  st_drop_geometry()%>%
  arrange(tag_id,t)%>%
  group_by(tag_id)%>%
  ungroup()%>%
  as.data.frame()%>%
  rename(ID=tag_id)

#prepare data and calculate step lengths and turning angles 
MS_filter<- MS%>%
  group_by(ID)%>%
  mutate(n=n())%>%
  filter(n>50)%>%
  ungroup()%>%as.data.frame()

data_hmm_MS <- prepData(data = MS_filter) %>%
  dplyr::mutate(across(
    .cols = c(dist_kettlehole:dist_river),
    .fns = ~ as.numeric(scale(.x))
  ))

MS_2024_X<-fitHMM(data_hmm_MS,
                nbStates=nbStates,
                dist=list(step=stepDist,angle=angleDist),
                Par0=Par0_M1,
                #workBounds=list(step=stepworkBounds,angle=angleworkBounds),
                #userBounds=list(step=stepBounds,angle=angleBounds),
                stateNames=stateNames,
                estAngleMean = list(angle=TRUE),
              formula = ~ dist_kettlehole*blood_infection,
              formulaDelta = ~  blood_infection)

saveRDS(MS_2024_X,here("output",
                    "data-proc",
                    "MS_2024_30min_X.RDS"))


plotStationary(MS_2024_X,plotCI = TRUE)

viterbi_MS_1<-momentuHMM::viterbi(MS_2024_X)

plot(MS_2024_X,plotTracks = T)

### get the states in

library("amt")


Ms_states<-MS_filter%>%
  mutate(state=viterbi_MS_1)%>%
  filter(state!=1)%>%
  #make amt track
  make_track(x,y,TIME,
             crs = 3035,
             all_cols=TRUE,
             id=ID)%>%
  nest(data= -c(ID,blood_infection))%>%
  mutate(data = map(data, ~steps(.x,keep_cols="end")),
         data = map(data, ~random_steps(.x,ncontrol=5)),
         data = map(data, ~extract_covariates(.x,ras)),
         data = map(data, ~mutate(.x,
                                  log_sl_ = log(sl_),
                                  sin_ta_ = sin(ta_),
                                  step_id1_=paste(ID,step_id_,sep="."),
                                  landuse_reclass=as.factor(landuse_reclass))))%>%
  unnest(cols = data)%>%
  mutate(human1=ifelse(landuse_reclass==1,1,0),
         agriculture1=ifelse(landuse_reclass==2,1,0),
         natural1=ifelse(landuse_reclass==3,1,0),
         water1=ifelse(landuse_reclass==4,1,0))


MS_m_30 <-
  glmmTMB::glmmTMB(
    case_ ~ -1 + agriculture1 * blood_infection + natural1 * blood_infection + water1 * blood_infection + dist_human * blood_infection +
      sl_ +  (1 |
                ID) + (0 + agriculture1 |
                             ID) + (0 + natural1 | ID) + (0 + water1 | ID),
    family = poisson,
    data = Ms_states,
    doFit = FALSE
  )

MS_m_30$theta[1] <- log(1e3)
MS_m_30$mapArg <-list(theta=factor(c(NA,1:3)))

MS_m4<-glmmTMB::fitTMB(MS_m_30)

summary(MS_m4)
####


tidy_MS_m4<-broom.mixed::tidy(MS_m4)%>%
  filter(effect=="fixed")%>%
  mutate(habitat = case_when(
    str_detect(term, "agriculture") ~ "agriculture",
    str_detect(term, "natural") ~ "semi-natural",
    str_detect(term, "water") ~ "water",
    TRUE ~ "other"  # default case 
  ))%>%
  filter(habitat!="other")%>%
  mutate(blood_infection = case_when(
    str_detect(term, "infectiony") ~ "yes",
    TRUE ~ "no"  # default case 
  ))



tidy_MS_m4%>%
  ggplot(aes(x=habitat,y=estimate,colour=blood_infection,group=blood_infection))+
  geom_pointrange(aes(ymin = estimate-1.96*std.error, ymax = estimate+1.96*std.error),position = position_jitterdodge(jitter.width = 0.05),size=2,linewidth=3)+
  d6::theme_d6(base_size = 30)+
  ylim(-1,1)+
  labs(y="Relative selection strength")+
  geom_hline(yintercept = 0,lty="dashed",colour="grey20")+
      scale_colour_manual(values = c("#5BBCD6","#F98400"),
                    labels=c("non-infected","infected"),
                    name="Infection status")





```

# 10 min
```{r}
#ctmm 10 min


swallows<-left_join(swallows_10min,cmr,by="tag_id")%>%
  mutate(blood_infection=as.factor(blood_infection))%>%
  mutate(date=as_date(TIME))%>%
  unite("ID",tag_id,date,remove=F)


ctmm_prep<-swallows%>%
  sf::st_drop_geometry() %>%
  # filter out all animals that have not been blood-sampled
  filter(!is.na(blood_infection)) %>%
  #rename
  mutate(
    sensor = "atlas",
    tag_id = as.character(tag_id),
    TAG = tag_id) %>%
  #doesn't like tibbles
  as.data.frame() %>%
  #rename
  rename(
    individual.local.identifier = tag_id,
    timestamp = TIME,
    location.long = long,
    location.lat = lat)%>%
  #telemetry object
  as.telemetry(
    projection = "+proj=laea +lat_0=52 +lon_0=10 +x_0=4321000 +y_0=3210000 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs +type=crs",keep = TRUE)


uere(ctmm_prep)<-1


# set-up parallel sessions
cl <- parallel::makeCluster(4)
doParallel::registerDoParallel(cl)

# create function to fit data & select best movement model
model_fit <- function(i)
{
  GUESS <-
    ctmm.guess(ctmm_prep[[i]],
               CTMM = ctmm(error = TRUE),
               interactive = FALSE)
  ctmm.select(ctmm_prep[[i]], GUESS)
}

#iterate over list
ctmm_fit_10min <-
  foreach(
    i = 1:length(ctmm_prep),
    .packages = 'ctmm',
    .errorhandling = 'remove'
  ) %dopar% {
    model_fit(i)
  }

#same names
names(ctmm_fit_10min) <- names(ctmm_prep)

# save
saveRDS(ctmm_fit_10min,
        here("output",
             "data-proc",
             "ctmm_fit_10min.RDS"))


# function for calculating akde for each individual (using the ctmm model)
akde_fit_weights <- function(i)
{
  AKDE <- akde(ctmm_prep[[i]],
               ctmm_fit_10min[[i]],
               weights = F)
}

#iterate over list
AKDES_tag_weights <- foreach(i=1: length(ctmm_prep),.packages='ctmm',.errorhandling = 'remove') %dopar% { akde_fit_weights(i) }

names(AKDES_tag_weights)<-names(ctmm_prep)

saveRDS(AKDES_tag_weights,here("output",
                          "data-proc",
                          "akdes_tag_10min.RDS"))

AKDE_size<-NULL

for(i in 1:length(AKDES_tag_weights)){
  TAG <- names(AKDES_tag_weights[i])
  min_AREA <- summary(AKDES_tag_weights[[i]],units=F)$CI[1]
  mean_AREA <- summary(AKDES_tag_weights[[i]],units=F)$CI[2]
  max_AREA <- summary(AKDES_tag_weights[[i]],units=F)$CI[3]
  blood_infection <- ctmm_prep[[i]]$blood_infection[1]
  species <- ctmm_prep[[i]]$species[1]
  
  AKDE_size = rbind(AKDE_size, data.frame(TAG,min_AREA,mean_AREA,max_AREA,blood_infection,species))
  
}


healthy_barn <-  AKDE_size %>%
    mutate(index = row_number())%>%
  filter(species == "RS" & blood_infection == "n")%>%
  pull(index)

infected_barn <-  AKDE_size %>%
    mutate(index = row_number())%>%
  filter(species == "RS" & blood_infection == "y")%>%
  pull(index)

# meta analysis
# akde 95%
meta(list(BS_NEG=
            AKDES_tag_weights[healthy_barn],
          BS_POS=
            AKDES_tag_weights[infected_barn]),
      units=TRUE,
     plot = TRUE,
     verbose = TRUE,
     level.UD = 0.95,
     level = 0.95,debias = F,
     main="homerange 95% AKDE isopleth")


healthy_house <- AKDE_size %>%
    mutate(index = row_number())%>%
  filter(species == "MS" & blood_infection == "n")%>%
  pull(index)

infected_house <- AKDE_size %>%
  mutate(index = row_number()) %>%
  filter(species == "MS" & blood_infection == "y") %>%
  pull(index)

# meta analysis
meta(list(HM_NEG=
            AKDES_tag_weights[healthy_house],
          HM_POS=
            AKDES_tag_weights[infected_house]),
     plot = TRUE,
     verbose = TRUE,
     level.UD = 0.95,
     level = 0.95,debias = F,
     main="homerange 95% AKDE isopleth")

```

## HMM
```{r}

swallows_10min<-left_join(swallows_10min,cmr,by="tag_id")%>%
  mutate(blood_infection=as.factor(blood_infection))%>%
  mutate(date=as_date(TIME))

ras <- terra::rast(here("output",
                        "geo-proc",
                        "stacked_2.tif"))


covariates<-terra::extract(ras,swallows_10min)%>%
  st_drop_geometry()%>%
  select(-ID)%>%
  as.data.frame()


swallows_cov<-bind_cols(swallows_10min,covariates)

RS <- swallows_cov %>% 
  dplyr::filter(species == "RS") %>%
  st_drop_geometry()%>%
  arrange(tag_id,t)%>%
  group_by(tag_id)%>%
  ungroup()%>%
  as.data.frame()%>%
  rename(ID=tag_id)

#prepare data and calculate step lengths and turning angles 
RS_filter<- RS%>%
  group_by(ID)%>%
  mutate(n=n())%>%
  filter(n>50)%>%
  ungroup()%>%as.data.frame()

#prepare data and calculate step lengths and turning angles 

data_hmm_RS <- prepData(data = RS_filter)%>%
  dplyr::mutate(across(.cols = c(dist_kettlehole:dist_river), .fns = ~ as.numeric(scale(.x))))


hist(data_hmm_RS$step,breaks = 20)

hist(data_hmm_RS$angle,breaks = 20)

#Nr states
nbStates<-3
# names
stateNames = c("resting","foraging","commuting")
# distributions
stepDist<-"gamma"
angleDist<-"vm"


# working bounds for step length

stepworkBounds <- matrix(c(rep(-Inf,6),
                           rep(Inf,6)),
                         6,#rows
                         2) #cols

#working bounds for turning angles

angleworkBounds <-  matrix(c(rep(-Inf,6),rep(Inf,6)),6,2)

stepBounds<-matrix(c(0,20,
                     0,Inf,
                     0,Inf,#40kmh
                     0,INF,
                     0,Inf,
                     0,Inf),nrow=2*nbStates,byrow=TRUE)



angleBounds<-matrix(c(0,pi,
                      0,pi,
                      0,1,
                      0.01,5,
                      0.01,5,
                      2,10),nrow=2*nbStates,byrow=TRUE)


40 * 1000 / 3600 * 60 * 10



stepPar0 <- c(10,500,6000,
              5,200,1000
              #0.95,0.025,0.025
              )

anglePar0 <-c(3.1,3.1,0.01,1,1,1)

Par0_M1<-list(step=stepPar0,
         angle=anglePar0)




RS_2024_X<-fitHMM(data_hmm_RS,
                nbStates=nbStates,
                dist=list(step=stepDist,angle=angleDist),
                Par0=Par0_M1,
                #workBounds=list(step=stepworkBounds,angle=angleworkBounds),
                #userBounds=list(step=stepBounds,angle=angleBounds),
                stateNames=stateNames,
                estAngleMean = list(angle=TRUE),
              formula = ~ dist_kettlehole*blood_infection,
              formulaDelta = ~  blood_infection)



saveRDS(RS_2024_X,here("output",
                    "data-proc",
                    "RS_2024_10min_X.RDS"))



plotStationary(RS_2024_X,plotCI = TRUE)

viterbi_RS_1<-momentuHMM::viterbi(RS_2024_X)

plot(RS_2024_X,plotTracks = T,ask=F)



Rs_states<-RS_filter%>%
  mutate(state=viterbi_RS_1)%>%
  # not resting
  filter(state!=1)%>%
  #make amt track
  make_track(x,y,TIME,
             crs = 3035,
             all_cols=TRUE,
             id=ID)%>%
  nest(data= -c(ID,blood_infection))%>%
  mutate(data = map(data, ~steps(.x,keep_cols="end")),
         data = map(data, ~random_steps(.x,ncontrol=5)),
         data = map(data, ~extract_covariates(.x,ras)),
         data = map(data, ~mutate(.x,
                                  log_sl_ = log(sl_),
                                  sin_ta_ = sin(ta_),
                                  step_id1_=paste(ID,step_id_,sep="."),
                                  landuse_reclass=as.factor(landuse_reclass))))%>%
  unnest(cols = data)%>%
  mutate(human1=ifelse(landuse_reclass==1,1,0),
         agriculture1=ifelse(landuse_reclass==2,1,0),
         natural1=ifelse(landuse_reclass==3,1,0),
         water1=ifelse(landuse_reclass==4,1,0))


RS_m_30 <-
  glmmTMB::glmmTMB(
    case_ ~ -1 + agriculture1 * blood_infection + natural1 * blood_infection + water1 * blood_infection + dist_human * blood_infection +
      sl_ +  (1 |
                ID) + (0 + agriculture1 |
                             ID) + (0 + natural1 | ID) + (0 + water1 | ID),
    family = poisson,
    data = Ms_states,
    doFit = FALSE
  )

RS_m_30$theta[1] <- log(1e3)
RS_m_30$mapArg <-list(theta=factor(c(NA,1:3)))

RS_m4<-glmmTMB::fitTMB(RS_m_30)

summary(RS_m4)
####


tidy_RS_m4<-broom.mixed::tidy(RS_m4)%>%
  filter(effect=="fixed")%>%
  mutate(habitat = case_when(
    str_detect(term, "agriculture") ~ "agriculture",
    str_detect(term, "natural") ~ "semi-natural",
    str_detect(term, "water") ~ "water",
    TRUE ~ "other"  # default case 
  ))%>%
  filter(habitat!="other")%>%
  mutate(blood_infection = case_when(
    str_detect(term, "infectiony") ~ "yes",
    TRUE ~ "no"  # default case 
  ))



tidy_RS_m4%>%
  ggplot(aes(x=habitat,y=estimate,colour=blood_infection,group=blood_infection))+
  geom_pointrange(aes(ymin = estimate-1.96*std.error, ymax = estimate+1.96*std.error),position = position_jitterdodge(jitter.width = 0.05),size=2,linewidth=3)+
  d6::theme_d6(base_size = 30)+
  ylim(-1,1)+
  labs(y="Relative selection strength")+
  geom_hline(yintercept = 0,lty="dashed",colour="grey20")+
      scale_colour_manual(values = c("#5BBCD6","#F98400"),
                    labels=c("non-infected","infected"),
                    name="Infection status")

#############################

MS <- swallows_cov %>% 
  dplyr::filter(species == "MS") %>%
  st_drop_geometry()%>%
  arrange(tag_id,t)%>%
  group_by(tag_id)%>%
  ungroup()%>%
  as.data.frame()%>%
  rename(ID=tag_id)

#prepare data and calculate step lengths and turning angles 
MS_filter<- MS%>%
  group_by(ID)%>%
  mutate(n=n())%>%
  filter(n>50)%>%
  ungroup()%>%as.data.frame()

data_hmm_MS <- prepData(data = MS_filter) %>%
  dplyr::mutate(across(
    .cols = c(dist_kettlehole:dist_river),
    .fns = ~ as.numeric(scale(.x))
  ))

MS_2024_X<-fitHMM(data_hmm_MS,
                nbStates=nbStates,
                dist=list(step=stepDist,angle=angleDist),
                Par0=Par0_M1,
                #workBounds=list(step=stepworkBounds,angle=angleworkBounds),
                #userBounds=list(step=stepBounds,angle=angleBounds),
                stateNames=stateNames,
                estAngleMean = list(angle=TRUE),
              formula = ~ dist_kettlehole*blood_infection,
              formulaDelta = ~  blood_infection)

saveRDS(MS_2024_X,here("output",
                    "data-proc",
                    "MS_2024_30min_X.RDS"))


plotStationary(MS_2024_X,plotCI = TRUE)

viterbi_MS_1<-momentuHMM::viterbi(MS_2024_X)

plot(MS_2024_X,plotTracks = T)

### get the states in

library("amt")


Ms_states<-MS_filter%>%
  mutate(state=viterbi_MS_1)%>%
  filter(state!=1)%>%
  #make amt track
  make_track(x,y,TIME,
             crs = 3035,
             all_cols=TRUE,
             id=ID)%>%
  nest(data= -c(ID,blood_infection))%>%
  mutate(data = map(data, ~steps(.x,keep_cols="end")),
         data = map(data, ~random_steps(.x,ncontrol=5)),
         data = map(data, ~extract_covariates(.x,ras)),
         data = map(data, ~mutate(.x,
                                  log_sl_ = log(sl_),
                                  sin_ta_ = sin(ta_),
                                  step_id1_=paste(ID,step_id_,sep="."),
                                  landuse_reclass=as.factor(landuse_reclass))))%>%
  unnest(cols = data)%>%
  mutate(human1=ifelse(landuse_reclass==1,1,0),
         agriculture1=ifelse(landuse_reclass==2,1,0),
         natural1=ifelse(landuse_reclass==3,1,0),
         water1=ifelse(landuse_reclass==4,1,0))


MS_m_30 <-
  glmmTMB::glmmTMB(
    case_ ~ -1 + agriculture1 * blood_infection + natural1 * blood_infection + water1 * blood_infection + dist_human * blood_infection +
      sl_ +  (1 |
                ID) + (0 + agriculture1 |
                             ID) + (0 + natural1 | ID) + (0 + water1 | ID),
    family = poisson,
    data = Ms_states,
    doFit = FALSE
  )

MS_m_30$theta[1] <- log(1e3)
MS_m_30$mapArg <-list(theta=factor(c(NA,1:3)))

MS_m4<-glmmTMB::fitTMB(MS_m_30)

summary(MS_m4)
####


tidy_MS_m4<-broom.mixed::tidy(MS_m4)%>%
  filter(effect=="fixed")%>%
  mutate(habitat = case_when(
    str_detect(term, "agriculture") ~ "agriculture",
    str_detect(term, "natural") ~ "semi-natural",
    str_detect(term, "water") ~ "water",
    TRUE ~ "other"  # default case 
  ))%>%
  filter(habitat!="other")%>%
  mutate(blood_infection = case_when(
    str_detect(term, "infectiony") ~ "yes",
    TRUE ~ "no"  # default case 
  ))



tidy_MS_m4%>%
  ggplot(aes(x=habitat,y=estimate,colour=blood_infection,group=blood_infection))+
  geom_pointrange(aes(ymin = estimate-1.96*std.error, ymax = estimate+1.96*std.error),position = position_jitterdodge(jitter.width = 0.05),size=2,linewidth=3)+
  d6::theme_d6(base_size = 30)+
  ylim(-1,1)+
  labs(y="Relative selection strength")+
  geom_hline(yintercept = 0,lty="dashed",colour="grey20")+
      scale_colour_manual(values = c("#5BBCD6","#F98400"),
                    labels=c("non-infected","infected"),
                    name="Infection status")





```
